[REFACTOR] Move methods *FormatUtils->CoreUtils
